Store Nucleus A/Z as int, drop unused Matrix equality operators (PR #26 review)
Nucleus::A/Z and Glauber's currentA1_/currentA2_/currentZ1_/currentZ2_
were doubles even though every assignment site sets a whole-number
literal and every external caller already casts back to int via
nucleusA1/A2/Z1/Z2. This left Glauber::interNuPInSP/interNuTInST's
`== 1.0` checks as floating-point comparisons for no reason.

Also remove Matrix::operator==/operator!=: unused anywhere in the
codebase and did a direct floating-point equality comparison on
complex<double> elements.

+++ src/Glauber.h
@@ -23,8 +23,8 @@ struct ReturnValue {
 
 struct Nucleus {
     std::string name;
-    double A;
-    double Z;
+    int A;
+    int Z;
     int anumFunc;
     int anumFuncIntegrand;
     int densityFunc;
@@ -57,19 +57,19 @@ class Glauber {
     Nucleus *Nuc_WS_;
     Data glauberData_;
     double b_;  // impact parameter
-    double currentA1_;
-    double currentA2_;
-    double currentZ1_;
-    double currentZ2_;
+    int currentA1_;
+    int currentA2_;
+    int currentZ1_;
+    int currentZ2_;
 
   public:
     Glauber() {};
     ~Glauber() { remove("tmp.dat"); }
 
-    int nucleusA1() const { return static_cast<int>(currentA1_); }
-    int nucleusA2() const { return static_cast<int>(currentA2_); }
-    int nucleusZ1() const { return static_cast<int>(currentZ1_); }
-    int nucleusZ2() const { return static_cast<int>(currentZ2_); }
+    int nucleusA1() const { return currentA1_; }
+    int nucleusA2() const { return currentA2_; }
+    int nucleusZ1() const { return currentZ1_; }
+    int nucleusZ2() const { return currentZ2_; }
     const Data &getGlauberData() const { return glauberData_; }
     int isFile(char *file_name);
     void findNucleusData(
